Cardano.Crypto.Praos.VRF

Key

data SecretKey

toPublic :: SecretKey -> PublicKey

get the public key associated to the given secret key

data PublicKey

data KeyPair

Constructors

KeyPair

Fields

toPublicKey :: !PublicKey

toSecretKey :: !SecretKey

generateKeyPair :: MonadRandom randomly => randomly KeyPair

generate a new secret key

generate

generateKey :: MonadRandom randomly => randomly SecretKey

generate a new secret key

Serialisation

secretKeyToBytes :: ByteArray b => SecretKey -> b

serialise a secret key into a binary representation

publicKeyToBytes :: ByteArray b => PublicKey -> b

serialise a public key into a binary representation

secretKeyFromBytes :: ByteArrayAccess ba => ba -> SecretKey

get a secret key from the given bytes

publicKeyFromBytes :: ByteArrayAccess ba => ba -> Either LString PublicKey

get a public key from a binary representation

this function may fail due to the nature of the underlying types of a public key.

Proof

data Proof

Constructors

Proof PublicKey Proof

generate :: (ByteArrayAccess ba, MonadRandom randomly) => ba -> SecretKey -> randomly (Output, Proof)

Generate a Deterministicaly _random_ Output and an associated Proof

At any time, with the associated PublicKey one is able to verify the Output of generate with the returned Output and Proof.

verify :: ByteArrayAccess ba => ba -> PublicKey -> (Output, Proof) -> Bool

verify that the given message and the SecretKey associated to the given PublicKey generated the given Output.

for testing

generate' :: ByteArrayAccess ba => SecretKey -> ba -> SecretKey -> (Output, Proof)

Generate a Deterministicaly _random_ Output and an associated Proof

At any time, with the associated PublicKey one is able to verify the Output of generate with the returned Output and Proof.