Crypto.ECC.Ed25519BIP32

Documentation

type MasterSecret = FBits 256

A Master secret is a 256 bits random quantity

type ChildKey (didxs :: DerivationHier) = Key

A child key is similar to the key in structure except it has an additional annotation representing the indexes for the hierarchy derivation indexes from a base Key (usually the root key)

type Key = (FBits 256, FBits 256, ChainCode)

A key is a 512 bit random value and a chaincode

Left half need to have: * Lowest 3 bits clear * Highest bit clear * Second highest bit set * Third highest bit clear

Right half doesn't have any particular structure.

type Public = (PointCompressed, ChainCode)

A public part of a key

type PointCompressed = FBits 256

A point is 1 bit of x sign and 255 bit of y coordinate (y's 256th bit is always 0)

newtype ChainCode

A 256 bits chain code

Constructors

ChainCode

Fields

unChainCode :: Bytes 32

newtype Hash n

A n bits Digest

Constructors

Hash

Fields

unHash :: FBits n

type Tag = Bytes 1

A Serialized tag used during HMAC

newtype SerializedIndex

Serialized Index

Constructors

SerializedIndex (Bytes 4)

type HMAC_SHA512 = Bytes 64

data DerivationType

Constructors

Hard

Soft

data DerivationMaterial

Constructors

ChainCodeMaterial

KeyMaterial

data DerivationIndex (k :: DerivationType) (n :: Nat)

Constructors

DerivationIndex

data DerivationHier

Constructors

(:>) Nat DerivationHier

DerivationEnd

type MaxHardIndex = 4294967295

type MinHardIndex = 2147483648

type MaxSoftIndex = MinHardIndex - 1

type MinSoftIndex = 0

data ValidIndex :: Nat -> Type where

Constructors

IsValidIndex :: (ValidDerivationIndex n :~: 'True) -> ValidIndex n

IsNotValidIndex :: (ValidDerivationIndex n :~: 'False) -> ValidIndex n

data ValidHardIndex :: Nat -> Type where

Constructors

IsValidHardIndex :: (ValidDerivationHardIndex n :~: 'True) -> ValidHardIndex n

IsNotValidHardIndex :: (ValidDerivationHardIndex n :~: 'False) -> ValidHardIndex n

data ValidSoftIndex :: Nat -> Type where

Constructors

IsValidSoftIndex :: (ValidDerivationSoftIndex n :~: 'True) -> ValidSoftIndex n

IsNotValidSoftIndex :: (ValidDerivationSoftIndex n :~: 'False) -> ValidSoftIndex n

getValidIndex :: KnownNat n => Proxy n -> Maybe (ValidDerivationIndex n :~: 'True)

isValidIndex :: KnownNat n => Proxy n -> ValidIndex n

getValidHardIndex :: KnownNat n => Proxy n -> Maybe (ValidDerivationHardIndex n :~: 'True)

isValidHardIndex :: KnownNat n => Proxy n -> ValidHardIndex n

getValidSoftIndex :: KnownNat n => Proxy n -> Maybe (ValidDerivationSoftIndex n :~: 'True)

isValidSoftIndex :: KnownNat n => Proxy n -> ValidSoftIndex n

type ValidDerivationIndex (n :: Nat) = (MinSoftIndex <=? n) && (n <=? MaxHardIndex)

type ValidDerivationHardIndex (n :: Nat) = (MinHardIndex <=? n) && (n <=? MaxHardIndex)

type ValidDerivationSoftIndex (n :: Nat) = (MinSoftIndex <=? n) && (n <=? MaxSoftIndex)

type family ValidDerivationIndexForType (k :: DerivationType) (n :: Nat) :: Bool where ...

Equations

ValidDerivationIndexForType 'Hard n = ValidDerivationHardIndex n

ValidDerivationIndexForType 'Soft n = ValidDerivationSoftIndex n

type family DerivationTag (ty :: DerivationType) (material :: DerivationMaterial) :: Nat where ...

Equations

DerivationTag 'Hard 'KeyMaterial = 0

DerivationTag 'Hard 'ChainCodeMaterial = 1

DerivationTag 'Soft 'KeyMaterial = 2

DerivationTag 'Soft 'ChainCodeMaterial = 3

leftHalfValid :: FBits 256 -> Bool

Check if the left half is valid

toPublic :: Key -> Public

kToPoint :: FBits 256 -> PointCompressed

pointAdd :: PointCompressed -> PointCompressed -> PointCompressed

pointToRepr :: PointCompressed -> PointCompressed

pointFromRepr :: PointCompressed -> PointCompressed

type family BitsToHashScheme (n :: Nat) where ...

Equations

BitsToHashScheme 256 = SHA256

BitsToHashScheme 512 = SHA512

type ValidTag tag = (0 <= tag, tag <= 3)

fcp :: forall tag idx deriveType deriveMaterial. (KnownNat (DerivationTag deriveType deriveMaterial), KnownNat idx, DerivationTag deriveType deriveMaterial ~ tag, ValidDerivationIndex idx ~ 'True, ValidDerivationIndexForType deriveType idx ~ 'True) => Proxy deriveMaterial -> Proxy deriveType -> Proxy idx -> ChainCode -> DerivationIndex deriveType idx -> [Word8] -> HMAC_SHA512

Compute the HMAC-SHA512 using the ChainCode as the key

hmacSHA512 :: Bytes keyLength -> Bytes input -> HMAC_SHA512

class GetDerivationMaterial (dtype :: DerivationType) mat where

Methods

getDerivationMaterial :: Proxy dtype -> mat -> [Word8]

derive :: forall dtype idx. (KnownNat (DerivationTag dtype 'KeyMaterial), KnownNat (DerivationTag dtype 'ChainCodeMaterial), KnownNat idx, ValidDerivationIndex idx ~ 'True, ValidDerivationIndexForType dtype idx ~ 'True, GetDerivationMaterial dtype Key) => DerivationIndex dtype idx -> Key -> Key

derivePublic :: forall idx dtype. (dtype ~ 'Soft, KnownNat (DerivationTag dtype 'KeyMaterial), KnownNat (DerivationTag dtype 'ChainCodeMaterial), KnownNat idx, ValidDerivationIndex idx ~ 'True, ValidDerivationIndexForType dtype idx ~ 'True, GetDerivationMaterial dtype PointCompressed) => DerivationIndex 'Soft idx -> PointCompressed -> ChainCode -> (PointCompressed, ChainCode)

step2 :: Bytes 64 -> (FBits 256, FBits 256)

Given Z, return 8*ZL(28Bytes) and ZR

indexSerialized :: forall idx. (KnownNat idx, ValidDerivationIndex idx ~ 'True) => Proxy idx -> SerializedIndex

Serialized index