Crypto.Encoding.BIP39.Dictionary

Dictionary

data Dictionary

this discribe the property of the Dictionary and will alllow to convert from a mnemonic phrase to MnemonicSentence

This is especially needed to build the BIP39 Seed

Constructors

Dictionary

Fields

dictionaryIndexToWord :: WordIndex -> String

This function will retrieve the mnemonic word associated to the given WordIndex.

dictionaryWordToIndex :: String -> Either DictionaryError WordIndex

This function will retrieve the WordIndex from a given mnemonic word.

dictionaryTestWord :: String -> Bool

test a given word is in the dictionary

dictionaryWordSeparator :: String

joining string (e.g. space for english)

data WordIndex

Index of the mnemonic word in the Dictionary

WordIndex are within range of [0..2047]

wordIndex :: Offset String -> WordIndex

unWordIndex :: WordIndex -> Offset String

data DictionaryError

Constructors

ErrInvalidDictionaryWord String