Crypto.Math.Edwards25519

Description

Simple module to play with the arithmetics of the twisted edwards curve Ed25519 using Extended Twisted Edwards Coordinates. Compared to the normal implementation this allow to use standard DH property:

for all valid s1 and s2 scalar:

scalarToPoint (s1 + s2) = pointAdd (scalarToPoint s1) (scalarToPoint s2)

For further useful references about Ed25519:

RFC 8032

http://ed25519.cr.yp.to/

http://ed25519.cr.yp.to/ed25519-20110926.pdf

http://eprint.iacr.org/2008/522.pdf

Basic types

data Scalar

Represent a scalar in the base field

data PointCompressed

Represent a point on the Edwards 25519 curve

newtype Signature

Represent a signature

Constructors

Signature

Fields

unSignature :: ByteString

smart constructor & destructor

scalar :: ByteString -> Scalar

Create a Ed25519 scalar

Only check that the length is of expected size (32 bytes), no effort is made for the scalar to be in the right base field range on purpose.

scalarP :: Bytes 32 -> Scalar

unScalar :: Scalar -> ByteString

pointCompressed :: HasCallStack => ByteString -> PointCompressed

Check if a scalar is valid and all the bits properly set/cleared scalarValid :: Scalar -> Bool scalarValid _s = True -- TODO

Smart constructor to create a compress point binary

Check if the length is of expected size

pointCompressedP :: Bytes 32 -> PointCompressed

unPointCompressed :: PointCompressed -> ByteString

unPointCompressedP :: PointCompressed -> Bytes 32

Arithmetic

scalarFromInteger :: Integer -> Scalar

Create a scalar from integer. mainly for debugging purpose.

scalarAdd :: Scalar -> Scalar -> Scalar

Add 2 scalar in the base field together

scalarToPoint :: Scalar -> PointCompressed

Lift a scalar to the curve, and returning a compressed point

pointAdd :: PointCompressed -> PointCompressed -> PointCompressed

Add 2 points together

Signature & Verify

sign :: ByteArrayAccess msg => Scalar -> ByteString -> msg -> Signature

Create a signature using a variant of ED25519 signature

we don't hash the secret key to derive a key + prefix, but instead we take an explicit salt and compute a prefix using the secret key + salt.

verify :: ByteArrayAccess msg => PointCompressed -> msg -> Signature -> Bool

Verify a signature