Cardano.Crypto.Wallet

Description

This provides similar functionality to BIP32 but using Ed25519 arithmetic instead of P256K1 arithmethic.

Key can be hierarchically derived from private key in two fashions: Hardened or Normal.

In the hardened scheme, the child secret key is not linearly derived, so that the child public key has no way to be efficiently computed from the parent public key.

The normal scheme allows anyone to derive public keys from public key.

Documentation

newtype ChainCode

Constructors

ChainCode ByteString

data DerivationScheme

Constructors

DerivationScheme1

DerivationScheme2

type DerivationIndex = Word32

pattern LatestScheme :: DerivationScheme

Extended Private & Public types

data XPrv

data XPub

Constructors

XPub

Fields

xpubPublicKey :: !ByteString

xpubChaincode :: !ChainCode

data XSignature

generate :: (ByteArrayAccess passPhrase, ByteArrayAccess seed) => seed -> passPhrase -> XPrv

Generate a new XPrv

The seed needs to be at least 32 bytes, otherwise an asynchronous error is thrown

generateNew

Arguments

:: (ByteArrayAccess keyPassPhrase, ByteArrayAccess generationPassPhrase, ByteArrayAccess seed)

=> seed
Raw entropy used as source of randomness for this algorithm

-> generationPassPhrase
User chosen passphrase for the generation phase

-> keyPassPhrase
Symmetric encryption key passphrase used for the in-memory security

-> XPrv

Generate a new XPrv from an entropy seed

The seed should be at least 16 bytes, although it is not enforced

The passphrase encrypt the secret key in memory

xprv :: ByteArrayAccess bin => bin -> Either String XPrv

Simple constructor

xpub :: ByteString -> Either String XPub

xsignature :: ByteString -> Either String XSignature

unXPrv :: XPrv -> ByteString

unXPub :: XPub -> ByteString

unXSignature :: XSignature -> ByteString

toXPub :: HasCallStack => XPrv -> XPub

Generate extended public key from private key

xPubGetPublicKey :: XPub -> PublicKey

Return the Ed25519 public key associated with a XPub context

xPrvChangePass

Arguments

:: (ByteArrayAccess oldPassPhrase, ByteArrayAccess newPassPhrase)

=> oldPassPhrase
passphrase to decrypt the current encrypted key

-> newPassPhrase
new passphrase to use for the new encrypted key

-> XPrv

-> XPrv

Derivation function

deriveXPrv :: ByteArrayAccess passPhrase => DerivationScheme -> passPhrase -> XPrv -> Word32 -> XPrv

Derive a child extended private key from an extended private key

deriveXPub :: DerivationScheme -> XPub -> Word32 -> Maybe XPub

Derive a child extended public key from an extended public key

Signature & Verification from extended types

sign :: (ByteArrayAccess passPhrase, ByteArrayAccess msg) => passPhrase -> XPrv -> msg -> XSignature

verify :: ByteArrayAccess msg => XPub -> msg -> XSignature -> Bool