Crypto.ECC.Ed25519Donna

Description

Ed25519 support

Documentation

newtype SecretKey

An Ed25519 Secret key

Constructors

SecretKey ScrubbedBytes

newtype PublicKey

An Ed25519 public key

Constructors

PublicKey Bytes

data Signature

An Ed25519 signature

Smart constructors

signature :: ByteArrayAccess ba => ba -> CryptoFailable Signature

Try to build a signature from a bytearray

publicKey :: ByteArrayAccess ba => ba -> CryptoFailable PublicKey

Try to build a public key from a bytearray

secretKey :: ByteArrayAccess ba => ba -> CryptoFailable SecretKey

Try to build a secret key from a bytearray

methods

toPublic :: SecretKey -> PublicKey

Create a public key from a secret key

sign :: (ByteArrayAccess msg, ByteArrayAccess salt) => SecretKey -> salt -> PublicKey -> msg -> Signature

Sign a message using the key pair

verify :: ByteArrayAccess ba => PublicKey -> ba -> Signature -> Bool

Verify a message

publicAdd :: PublicKey -> PublicKey -> PublicKey

secretAdd :: SecretKey -> SecretKey -> SecretKey