Crypto.Encoding.BIP39

Entropy

data Entropy (n :: Nat)

BIP39's entropy is a byte array of a given size (in bits, see ValidEntropySize for the valid size).

To it is associated

type ValidEntropySize (n :: Nat) = (KnownNat n, NatWithinBound Int n, Elem n '[96, 128, 160, 192, 224, 256])

Type Constraint Alias to check a given Nat is valid for an entropy size

i.e. it must be one of the following: 96, 128, 160, 192, 224, 256.

data Checksum (bits :: Nat)

this is the Checksum of a given Entropy

the Nat type parameter represent the size, in bits, of this checksum.

type ValidChecksumSize (ent :: Nat) (csz :: Nat) = (KnownNat csz, NatWithinBound Int csz, Elem csz '[3, 4, 5, 6, 7, 8], CheckSumBits ent ~ csz)

type family MnemonicWords (n :: Nat) :: Nat where ...

Number of Words related to a specific entropy size in bits

Equations

MnemonicWords 96 = 9

MnemonicWords 128 = 12

MnemonicWords 160 = 15

MnemonicWords 192 = 18

MnemonicWords 224 = 21

MnemonicWords 256 = 24

type family EntropySize (n :: Nat) :: Nat where ...

Corresponding entropy size in bits for a given number of words

Equations

EntropySize 9 = 96

EntropySize 12 = 128

EntropySize 15 = 160

EntropySize 18 = 192

EntropySize 21 = 224

EntropySize 24 = 256

toEntropy :: forall n csz ba. (ValidEntropySize n, ValidChecksumSize n csz, ByteArrayAccess ba) => ba -> Either (EntropyError csz) (Entropy n)

Create a specific entropy type of known size from a raw bytestring

entropyRaw :: Entropy n -> ByteString

Get the raw binary associated with the entropy

entropyChecksum :: Entropy n -> Checksum (CheckSumBits n)

Get the checksum of the Entropy

entropyToWords :: forall n csz mw. ConsistentEntropy n mw csz => Entropy n -> MnemonicSentence mw

Given an entropy of size n, Create a list

wordsToEntropy :: forall ent csz mw. ConsistentEntropy ent mw csz => MnemonicSentence mw -> Either (EntropyError csz) (Entropy ent)

retrieve the initial entropy from a given MnemonicSentence

This function validate the retrieved Entropy is valid, i.e. that the checksum is correct. This means you should not create a new Entropy from a MnemonicSentence, instead, you should use a Random Number Generator to create a new Entropy.

Seed

data Seed

type Passphrase = String

sentenceToSeed

Arguments

:: ValidMnemonicSentence mw

=> MnemonicSentence mw
MmenomicPhrase of mw words

-> Dictionary
Dictionary' of words/indexes

-> Passphrase
Passphrase used to generate

-> Seed

Create a seed from MmemonicSentence and Passphrase using the BIP39 algorithm.

phraseToSeed

Arguments

:: ValidMnemonicSentence mw

=> MnemonicPhrase mw
MmenomicPhrase of mw words

-> Dictionary
Dictionary' of words/indexes

-> Passphrase
Passphrase used to generate

-> Seed

Create a seed from MmemonicPhrase and Passphrase using the BIP39 algorithm.

Mnemonic Sentence

data MnemonicSentence (mw :: Nat)

Mnemonic Sentence is a list of WordIndex.

This is the generic representation of a mnemonic phrase that can be used for transalating to a different dictionary (example: English to Japanese).

This is mainly used to convert from/to the Entropy and for cardanoSlSeed

data MnemonicPhrase (mw :: Nat)

Human readable representation of a MnemonicSentence

type ValidMnemonicSentence (mw :: Nat) = (KnownNat mw, NatWithinBound Int mw, Elem mw '[9, 12, 15, 18, 21, 24])

Type Constraint to validate the given Nat is valid for the supported MnemonicSentence

mnemonicPhrase :: forall mw. ValidMnemonicSentence mw => [String] -> Either MnemonicWordsError (MnemonicPhrase mw)

checkMnemonicPhrase :: forall mw. ValidMnemonicSentence mw => Dictionary -> MnemonicPhrase mw -> Bool

check a given MnemonicPhrase is valid for the given Dictionary

mnemonicPhraseToMnemonicSentence :: forall mw. ValidMnemonicSentence mw => Dictionary -> MnemonicPhrase mw -> Either DictionaryError (MnemonicSentence mw)

convert the given MnemonicPhrase to a generic MnemonicSentence with the given Dictionary.

This function assumes the Dictionary and the MnemonicPhrase are compatible (see checkMnemonicPhrase).

mnemonicSentenceToMnemonicPhrase :: forall mw. ValidMnemonicSentence mw => Dictionary -> MnemonicSentence mw -> MnemonicPhrase mw

convert the given generic MnemonicSentence to a human readable MnemonicPhrase targetting the language of the given Dictionary.

mnemonicSentenceToString :: forall mw. ValidMnemonicSentence mw => Dictionary -> MnemonicSentence mw -> String

mnemonicSentenceToListN :: MnemonicSentence mw -> ListN mw WordIndex

mnemonicPhraseToString :: forall mw. ValidMnemonicSentence mw => Dictionary -> MnemonicPhrase mw -> String

translateTo

Arguments

:: forall mw. ValidMnemonicSentence mw

=> Dictionary
source dictionary

-> Dictionary
destination dictionary

-> MnemonicPhrase mw

-> Either DictionaryError (MnemonicPhrase mw)

translate the given MnemonicPhrase from one dictionary into another.

This function assumes the source dictionary is compatible with the given MnemonicPhrase (see checkMnemonicPhrase)

Dictionary

data Dictionary

this discribe the property of the Dictionary and will alllow to convert from a mnemonic phrase to MnemonicSentence

This is especially needed to build the BIP39 Seed

Constructors

Dictionary

Fields

dictionaryIndexToWord :: WordIndex -> String

This function will retrieve the mnemonic word associated to the given WordIndex.

dictionaryWordToIndex :: String -> Either DictionaryError WordIndex

This function will retrieve the WordIndex from a given mnemonic word.

dictionaryTestWord :: String -> Bool

test a given word is in the dictionary

dictionaryWordSeparator :: String

joining string (e.g. space for english)

data WordIndex

Index of the mnemonic word in the Dictionary

WordIndex are within range of [0..2047]

wordIndex :: Offset String -> WordIndex

unWordIndex :: WordIndex -> Offset String

helpers

type ConsistentEntropy ent mw csz = (ValidEntropySize ent, ValidChecksumSize ent csz, ValidMnemonicSentence mw, MnemonicWords ent ~ mw)

Type Constraint Alias to check the entropy size, the number of mnemonic words and the checksum size is consistent. i.e. that the following is true:

| entropysize | checksumsize | entropysize + checksumsize | mnemonicsize | +---------------+--------------+----------------------------+--------------+ | 96 | 3 | 99 | 9 | | 128 | 4 | 132 | 12 | | 160 | 5 | 165 | 15 | | 192 | 6 | 198 | 18 | | 224 | 7 | 231 | 21 | | 256 | 8 | 264 | 24 |

This type constraint alias also perform all the GHC's cumbersome type level literal handling.

type family CheckSumBits (n :: Nat) :: Nat where ...

Number of bits of checksum related to a specific entropy size in bits

Equations

CheckSumBits 96 = 3

CheckSumBits 128 = 4

CheckSumBits 160 = 5

CheckSumBits 192 = 6

CheckSumBits 224 = 7

CheckSumBits 256 = 8

type family Elem (e :: Nat) (l :: [Nat]) :: Constraint where ...

convenient type level constraint to validate a given Nat e is an elemnt of the list of Nat l.

Equations

Elem e '[] = TypeError (('Text "offset: field " :<>: 'ShowType e) :<>: 'Text " not elements of valids values")

Elem e (e ': _) = ()

Elem e (_ ': xs) = Elem e xs

Errors

data DictionaryError

Constructors

ErrInvalidDictionaryWord String

data EntropyError csz

Constructors

ErrInvalidEntropyLength Int Int

ErrInvalidEntropyChecksum (Checksum csz) (Checksum csz)

data MnemonicWordsError

Constructors

ErrWrongNumberOfWords Int Int